add fluxo comentario

+++ src/BlogCore/Business/Models/Post.cs
@@ -4,7 +4,7 @@
     {
         public string? Titulo { get; set; }
         public string? Conteudo { get; set; }
-        public int AutorId { get; set; }
+        public long AutorId { get; set; }
         public required Autor Autor { get; set; }
         public ICollection<Comentario>? Comentarios { get; set; }
     }
